feat: splits compilation to two projects; outer cannot yet consume inner
