+++ src/tsconfig.module.json
@@ -1,20 +1,17 @@
 {
   "compilerOptions": {
     "allowJs": true,
-    "module": "none",
-    "outDir": "../dist/",
+    "composite": true,
+    "module": "commonjs",
+    "outDir": "../dist",
     "inlineSources": true,
     "sourceMap": true,
     "declaration": true,
     "target": "es5",
     "types": ["node"],
     "lib": ["es6"],
-    "outFile": "../dist/index.js"
   },
-  "files": [
-    "index.ts"
-  ],
-  "references": [
-    {"path": "./tsconfig.module.json" }
+  "include": [
+    "**/*.m.ts"
   ]
 }
