--- src/utils/index.ts
@@ -1,2 +1,0 @@
-///<reference path="deviceSpec.ts" />
-///<reference path="version.ts" />
